move inject_connection_established to PM's NB impl
