+++ beacon_node/lighthouse_network/src/peer_manager/network_behaviour.rs
@@ -7,11 +7,14 @@ use libp2p::swarm::{
     DialError, NetworkBehaviour, NetworkBehaviourAction, PollParameters, ProtocolsHandler,
 };
 use libp2p::{Multiaddr, PeerId};
+use slog::{debug, error};
 use types::EthSpec;
 
-use crate::PeerManager;
+use crate::metrics;
+use crate::rpc::GoodbyeReason;
 
-use super::PeerManagerEvent;
+use super::peerdb::BanResult;
+use super::{PeerManager, PeerManagerEvent, ReportSource};
 
 impl<TSpec: EthSpec> NetworkBehaviour for PeerManager<TSpec> {
     type ProtocolsHandler = DummyProtocolsHandler;
@@ -43,14 +46,6 @@ impl<TSpec: EthSpec> NetworkBehaviour for PeerManager<TSpec> {
 
     /* Overwritten trait members */
 
-    fn addresses_of_peer(&mut self, peer_id: &PeerId) -> Vec<Multiaddr> {
-        vec![]
-    }
-
-    fn inject_connected(&mut self, peer_id: &PeerId) {}
-
-    fn inject_disconnected(&mut self, peer_id: &PeerId) {}
-
     fn inject_connection_established(
         &mut self,
         peer_id: &PeerId,
@@ -58,16 +53,77 @@ impl<TSpec: EthSpec> NetworkBehaviour for PeerManager<TSpec> {
         endpoint: &ConnectedPoint,
         failed_addresses: Option<&Vec<Multiaddr>>,
     ) {
+        // Log the connection
+        match &endpoint {
+            ConnectedPoint::Listener { .. } => {
+                debug!(self.log, "Connection established"; "peer_id" => %peer_id, "connection" => "Incoming");
+            }
+            ConnectedPoint::Dialer { .. } => {
+                debug!(self.log, "Connection established"; "peer_id" => %peer_id, "connection" => "Outgoing");
+                // TODO: Ensure we have that address registered.
+            }
+        }
+
+        // Check to make sure the peer is not supposed to be banned
+        match self.ban_status(&peer_id) {
+            // TODO: directly emit the ban event?
+            BanResult::BadScore => {
+                // This is a faulty state
+                error!(self.log, "Connecteded to a banned peer, re-banning"; "peer_id" => %peer_id);
+                // Reban the peer
+                self.goodbye_peer(&peer_id, GoodbyeReason::Banned, ReportSource::PeerManager);
+                return;
+            }
+            BanResult::BannedIp(ip_addr) => {
+                // A good peer has connected to us via a banned IP address. We ban the peer and
+                // prevent future connections.
+                debug!(self.log, "Peer connected via banned IP. Banning"; "peer_id" => %peer_id, "banned_ip" => %ip_addr);
+                self.goodbye_peer(&peer_id, GoodbyeReason::BannedIP, ReportSource::PeerManager);
+                return;
+            }
+            BanResult::NotBanned => {}
+        }
+
+        // Check the connection limits
+        if self.peer_limit_reached()
+            && self
+                .network_globals
+                .peers
+                .read()
+                .peer_info(&peer_id)
+                .map_or(true, |peer| !peer.has_future_duty())
+        {
+            // Gracefully disconnect the peer.
+            self.disconnect_peer(*peer_id, GoodbyeReason::TooManyPeers);
+            return;
+        }
+
+        // Register the newly connected peer (regardless if we are about to disconnect them).
+        // NOTE: We don't register peers that we are disconnecting immediately. The network service
+        // does not need to know about these peers.
+        // let enr
+        match endpoint {
+            ConnectedPoint::Listener { send_back_addr, .. } => {
+                self.inject_connect_ingoing(&peer_id, send_back_addr.clone(), None);
+                self.events
+                    .push(PeerManagerEvent::PeerConnectedIncoming(*peer_id));
+            }
+            ConnectedPoint::Dialer { address } => {
+                self.inject_connect_outgoing(&peer_id, address.clone(), None);
+                self.events
+                    .push(PeerManagerEvent::PeerConnectedOutgoing(*peer_id));
+            }
+        }
+
+        let connected_peers = self.network_globals.connected_peers() as i64;
+
+        // increment prometheus metrics
+        metrics::inc_counter(&metrics::PEER_CONNECT_EVENT_COUNT);
+        metrics::set_gauge(&metrics::PEERS_CONNECTED, connected_peers);
+        metrics::set_gauge(&metrics::PEERS_CONNECTED_INTEROP, connected_peers);
     }
 
-    fn inject_connection_closed(
-        &mut self,
-        peer_id: &PeerId,
-        connection_id: &ConnectionId,
-        endpoint: &ConnectedPoint,
-        handler: DummyProtocolsHandler,
-    ) {
-    }
+    fn inject_disconnected(&mut self, peer_id: &PeerId) {}
 
     fn inject_address_change(
         &mut self,
