+++ beacon_node/lighthouse_network/src/service.rs
@@ -317,23 +317,11 @@ impl<TSpec: EthSpec> Service<TSpec> {
                     return Libp2pEvent::Behaviour(behaviour);
                 }
                 SwarmEvent::ConnectionEstablished {
-                    peer_id,
-                    endpoint,
-                    num_established,
+                    peer_id: _,
+                    endpoint: _,
+                    num_established: _,
                     concurrent_dial_errors: _,
-                } => {
-                    // Inform the peer manager.
-                    // We require the ENR to inject into the peer db, if it exists.
-                    let enr = self
-                        .swarm
-                        .behaviour_mut()
-                        .discovery_mut()
-                        .enr_of_peer(&peer_id);
-                    self.swarm
-                        .behaviour_mut()
-                        .peer_manager_mut()
-                        .inject_connection_established(peer_id, endpoint, num_established, enr);
-                }
+                } => {}
                 SwarmEvent::ConnectionClosed {
                     peer_id,
                     cause: _,
